refactor(agent): extract instruction builders to src/agent/
Moved buildSkillGenerateInstruction() and SKILL_GENERATE_TASK from
src/daemon/api/agent-runs.ts to src/agent/instruction-builders.ts.
This is agent-layer logic (data assembly for agent prompts), not
API-layer logic. Both the API route handler and the scheduler now
import from the canonical location.
